WEBAPP-290: Rename WohnenExtra to WohnenExtraPage
